resolved minor changes
