--- src/.babelrc
@@ -1,7 +1,0 @@
-{
-    "env": {
-      "test": {
-        "plugins": ["@babel/plugin-transform-modules-commonjs"]
-      }
-    }
-}
