+++ src/test/counter.test.js
@@ -40,3 +40,23 @@ describe('Pokemon Counter function Test', () => {
     expect(pokemonLength).toBe(6);
   });
 });
+
+
+const comments = [
+  {
+    username: 'Test',
+    comment: 'James',
+  },
+  {
+    username: 'Test2',
+    comment: 'Mike',
+  },
+];
+
+describe('Comment Counter function Test', () => {
+  test('should return comment array length of 2', () => {
+    const len = counter.comments(comments);
+    expect(len).toBe(2);
+    expect(len).not.toBe(3);
+  });
+});
